MAINT: Rename ``spatial`` module and make them private
