+++ scipy/spatial/_ckdtree.pyi
