+++ scipy/spatial/_ckdtree.pyx
