+++ scipy/spatial/_qhull.pyi
