+++ scipy/spatial/transform/_rotation.pyi
