+++ scipy/spatial/transform/_rotation.pyx
